fix(sound): guard signed integer overflow in DMC acc and X6502 cycle scaling (hotfix3 C-3)
Two independent fix sites for signed-integer overflow UB reachable via
Lua/cheat supplying extreme cycle counts:

1. src/sound.cpp:521 - FCEU_SoundCPUHook

   The inner `std::min<uint32>(-DMCacc, ...)` negates DMCacc
   unconditionally. DMCacc is `int32_t&` (sound.h:82 → apu.h:146 →
   `dmc_acc_`). When DMCacc == INT32_MIN, `-DMCacc` is C++17
   [conv.prom] signed overflow UB. UBSan flags it; poisoned value
   may then leak through `std::min`'s `<` comparison.

   Fix: lift the negation through int64_t then cast back to uint32.
   This is well-defined at the boundary (yields 2^31 exactly) and
   matches the magnitude the original code intended.

2. src/x6502.cpp:453 - X6502_RunDebug

   Pre-fix:
       if (PAL) cycles *= 15;
       else      cycles *= 16;
       _count += cycles;

   With cycles int32 and PAL/non-PAL multiplier, `cycles *= 16`
   overflows at cycles == 2^27. Lua's fceu11_lua_x6502_run can pass
   such values (cheat scripts, stress tests); the resulting
   poisoned `_count` then drives the per-instruction dispatch
   loop into miscounted NMI/IRQ.

   Fix: scale in int64, clamp to INT32_MAX, cast back. Cost: one
   extra imul + cmp + cmov per `X6502_RunDebug` call site — at
   ~26k calls/sec @ 60 FPS this is well below the §十.2 1% soft
   target and invisible against the inner `while(_count>0)`
   RdMem/dispatch overhead.

3. tests/core/apu_test.cpp - two new boundary tests

   - test_dmc_acc_negation_overflow: set DMCacc = INT32_MIN, run
     one FCEU_SoundCPUHook(1), assert DMCacc advances past the
     sentinel (proves the inner while terminates under UBSan).
   - test_cycles_scaling_overflow: feed X6502_Run(INT32_MAX/8)
     (well above 2^27 UB threshold), then emulate 1 frame; both
     must complete without crashing.

PLAN §四 C-3 - https://github.com/Laffinty/FCEUX11/blob/main/docs/FCEUX11-1.15_LTS-hotfix3-PLAN.md

Side note: PLAN §四 C-3 originally listed the cycle-scaling site as
src/sound.cpp:453-455. That is incorrect - the actual file is
src/x6502.cpp:453-455 (the entry of X6502_RunDebug). Diagnosed during
the C-3 exploration; this commit applies the fix to the correct file.

Out of scope (identified during exploration, deferred to §十六):
  - ppu_rendering.cpp:226 (`linestartts = ... * 48` similar pattern)
  - cpu.h:106 Cpu::add_cycles (`layout_.count -= c * 48`)
  - sound.cpp:501 (`fhcnt -= cycles*48` similar shape)

ref: C-3

+++ src/sound.cpp
@@ -518,7 +518,11 @@ void FCEU_SoundCPUHook(int cycles)
    /* Unbelievably ugly hack */
    if(FSettings.SndRate)
    {
-		const uint32 fudge = std::min<uint32>(-DMCacc, soundtsoffs + g_cpu.timestamp_ref());
+		// hotfix3 C-3: lift negation through int64 — direct `-DMCacc` is signed
+		// overflow UB when DMCacc == INT32_MIN. Cast back to uint32 once safe.
+		const uint32 fudge = std::min<uint32>(
+				static_cast<uint32>(-(int64_t)DMCacc),
+				soundtsoffs + g_cpu.timestamp_ref());
 		soundtsoffs -= fudge;
 		DoPCM();
 		soundtsoffs += fudge;
